requirements.txt
torch>=2.2.0
torch-geometric>=2.5.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
networkx>=3.1
matplotlib>=3.7.0
scikit-learn>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
kagglehub>=0.2.5
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
